LayoutTests/editing/deleting/delete-block-table.html

<!DOCTYPE html>
<html>
<body>
<p>This test checks that deleting into a table works properly. When deleting three times with the cursor after the character "a" in the "after" text after the table, the "a" should be deleted, as should the "o" of "buffalo" in the last table cell.</p>
<div contenteditable="true" id="root">
<div>Before</div>
  <table style="border:3px solid #aaa;">
    <tr>
      <td>
        Foo
      </td>
      <td>
        baz
      </td>
    </tr>
    <tr>
      <td>
        bar
      </td>
      <td>
        buffalo
      </td>
    </tr>
  </table>
<span id="start">after</span>
</div>
<script src=../editing.js></script>
<script>
function editingTest() {
    var selection = window.getSelection();
    var s = document.getElementById("start");
    selection.setPosition(s, 0);
    execMoveSelectionForwardByCharacterCommand();
    for (i = 0; i < 3; i++) {
        deleteCommand();
    }
}
runDumpAsTextEditingTest();
</script>
</body>
</html>
